Move ComposerPublisher to Compose
